+++ include/gaia/ecs/component.h
@@ -240,14 +240,14 @@ namespace gaia {
 
 		[[nodiscard]] inline uint64_t CalculateMatcherHash(std::span<const ComponentInfo*> infos) noexcept {
 			uint64_t hash = infos.empty() ? 0 : infos[0]->matcherHash;
-			for (uint32_t i = 1U; i < (uint32_t)infos.size(); ++i)
+			for (size_t i = 1; i < infos.size(); ++i)
 				hash = utils::combine_or(hash, infos[i]->matcherHash);
 			return hash;
 		}
 
 		[[nodiscard]] inline uint64_t CalculateLookupHash(std::span<const ComponentInfo*> infos) noexcept {
 			uint64_t hash = infos.empty() ? 0 : infos[0]->lookupHash;
-			for (uint32_t i = 1U; i < (uint32_t)infos.size(); ++i)
+			for (size_t i = 1; i < infos.size(); ++i)
 				hash = utils::hash_combine(hash, infos[i]->lookupHash);
 			return hash;
 		}
